Déplace la logique d'éligibilité dans `core`
… pour pouvoir la réutiliser dans l'API.
On va vouloir envoyer à Metabase les résultats d'éligibilité. On ne veut pas faire confiance au statut envoyé par le front, pour éviter les bidouillages.
L'API va recalculer le statut avec le même code que le front.
