+++ anssi-nis2-ui/src/Components/Simulateur/AiguilleVersEtapeResultat.tsx
@@ -1,9 +1,9 @@
-import { EtatQuestionnaire } from "../../questionnaire/reducerQuestionnaire.ts";
 import { EtapeResultatV2 } from "./EtapeResultatV2.tsx";
 import { ResultatEligibilite } from "../../../../commun/core/src/Domain/Simulateur/Regulation.definitions.ts";
 import { ReactElement } from "react";
-import { evalueEligibilite } from "../../questionnaire/specifications/evalueEligibilite.ts";
-import SpecificationsCompletes from "../../questionnaire/specifications/specifications-completes.csv?raw";
+import { evalueEligibilite } from "../../../../commun/core/src/Domain/Questionnaire/evalueEligibilite.ts";
+import SpecificationsCompletes from "../../../../commun/core/src/Domain/Questionnaire/specifications-completes.csv?raw";
+import { EtatQuestionnaire } from "anssi-nis2-core/src/Domain/Questionnaire/EtatQuestionnaire";
 
 export function AiguilleVersEtapeResultat(props: {
   reponses: EtatQuestionnaire;
