+++ anssi-nis2-ui/src/Components/Simulateur/EtapesRefacto/EtapeActivites.tsx
@@ -4,9 +4,9 @@ import BlocPrincipal from "../../BlocPrincipal.tsx";
 import { FormSimulateur } from "../Etapes";
 import { SecteurSimple } from "anssi-nis2-core/src/Domain/Simulateur/SecteurActivite.definitions.ts";
 import { EnsembleChamps } from "../Inputs/EnsembleChamps.tsx";
-import { libellesSecteursActivite } from "../../../References/LibellesSecteursActivite.ts";
+import { libellesSecteursActivite } from "../../../../../commun/core/src/Domain/Simulateur/LibellesSecteursActivite.ts";
 import { SousSecteurActivite } from "anssi-nis2-core/src/Domain/Simulateur/SousSecteurActivite.definitions.ts";
-import { libellesSousSecteursActivite } from "../../../References/LibellesSousSecteursActivite.ts";
+import { libellesSousSecteursActivite } from "../../../../../commun/core/src/Domain/Simulateur/LibellesSousSecteursActivite.ts";
 import { secteurDe } from "anssi-nis2-core/src/Domain/Simulateur/services/SousSecteurActivite/SousSecteurActivite.operations.ts";
 import { activitesParSecteurEtSousSecteur } from "anssi-nis2-core/src/Domain/Simulateur/Activite.operations.ts";
 import { libellesActivites } from "../../../References/LibellesActivites.ts";
