+++ anssi-nis2-ui/src/Components/Simulateur/EtapesRefacto/EtapeSousSecteursActivite.tsx
@@ -6,9 +6,9 @@ import {
   SecteurComposite,
 } from "anssi-nis2-core/src/Domain/Simulateur/SecteurActivite.definitions.ts";
 import Checkbox from "@codegouvfr/react-dsfr/Checkbox";
-import { libellesSecteursActivite } from "../../../References/LibellesSecteursActivite.ts";
+import { libellesSecteursActivite } from "../../../../../commun/core/src/Domain/Simulateur/LibellesSecteursActivite.ts";
 import { fabriqueTupleSecteurSousSecteurs } from "anssi-nis2-core/src/Domain/Simulateur/services/SecteurActivite/SecteurActivite.operations.ts";
-import { libellesSousSecteursActivite } from "../../../References/LibellesSousSecteursActivite.ts";
+import { libellesSousSecteursActivite } from "../../../../../commun/core/src/Domain/Simulateur/LibellesSousSecteursActivite.ts";
 import { useState } from "react";
 import { SousSecteurActivite } from "anssi-nis2-core/src/Domain/Simulateur/SousSecteurActivite.definitions.ts";
 import { PrecedentSuivant } from "../PrecedentSuivant.tsx";
