+++ anssi-nis2-ui/src/Components/Simulateur/EtapesRefacto/PersisteReponsesDuQuestionnaire.tsx
@@ -1,6 +1,6 @@
-import { EnvoieDonneesFormulaire } from "../../../Services/Simulateur/Operations/appelsApi";
-import { EtatQuestionnaire } from "../../../questionnaire/reducerQuestionnaire.ts";
 import { ReactNode, useEffect } from "react";
+import { EnvoieDonneesFormulaire } from "../../../Services/Simulateur/Operations/appelsApi";
+import { EtatQuestionnaire } from "anssi-nis2-core/src/Domain/Questionnaire/EtatQuestionnaire";
 
 export function PersisteReponsesDuQuestionnaire(props: {
   persistance: EnvoieDonneesFormulaire;
