+++ anssi-nis2-ui/src/questionnaire/selecteursQuestionnaire.ts
@@ -1,9 +1,9 @@
 import { SecteurSimple } from "anssi-nis2-core/src/Domain/Simulateur/SecteurActivite.definitions.ts";
 import { SousSecteurActivite } from "anssi-nis2-core/src/Domain/Simulateur/SousSecteurActivite.definitions.ts";
-import { estSousSecteurListe } from "../../../commun/core/src/Domain/Simulateur/services/SousSecteurActivite/SousSecteurActivite.predicats.ts";
+import { estSousSecteurListe } from "anssi-nis2-core/src/Domain/Simulateur/services/SousSecteurActivite/SousSecteurActivite.predicats.ts";
 import { non } from "../../../commun/utils/services/commun.predicats.ts";
 import { estUnSecteurAvecDesSousSecteurs } from "anssi-nis2-core/src/Domain/Simulateur/services/SecteurActivite/SecteurActivite.predicats.ts";
-import { EtatQuestionnaire } from "./reducerQuestionnaire.ts";
+import { EtatQuestionnaire } from "anssi-nis2-core/src/Domain/Questionnaire/EtatQuestionnaire";
 
 export const selectSecteursPourSaisieActivites = (
   etat: EtatQuestionnaire,
