+++ anssi-nis2-ui/test/questionnaire/reducerUndo.spec.ts
@@ -5,10 +5,8 @@ import {
   quiSupporteUndo,
   undo,
 } from "../../src/questionnaire/quiSupporteUndo";
-import {
-  etatParDefaut,
-  EtatQuestionnaire,
-} from "../../src/questionnaire/reducerQuestionnaire";
+import { etatParDefaut } from "../../src/questionnaire/reducerQuestionnaire";
+import { EtatQuestionnaire } from "anssi-nis2-core/src/Domain/Questionnaire/EtatQuestionnaire";
 
 const listeActions: Pick<ActionQuestionnaire | ActionUndo, "type">[] = [
   { type: "VIDE" },
