+++ anssi-nis2-ui/test/questionnaire/selecteursQuestionnaire.spec.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from "vitest";
 import {
   etatParDefaut,
-  EtatQuestionnaire,
   reducerQuestionnaire,
 } from "../../src/questionnaire/reducerQuestionnaire";
 import {
@@ -10,6 +9,7 @@ import {
   valideSousSecteursActivite,
 } from "../../src/questionnaire/actions";
 import { selectSecteursPourSaisieActivites } from "../../src/questionnaire/selecteursQuestionnaire";
+import { EtatQuestionnaire } from "anssi-nis2-core/src/Domain/Questionnaire/EtatQuestionnaire";
 
 describe("Les sélecteurs sur l'etat du questionnaire", () => {
   describe("le sélecteur qui retourne les secteurs/sous-secteurs dont on veut connaître les activités", () => {
