+++ commun/.prettierrc
@@ -1,0 +1,3 @@
+{
+  "trailingComma": "all"
+}
