+++ commun/core/src/Domain/Questionnaire/EtatQuestionnaire.ts
@@ -1,0 +1,49 @@
+import { TypeEtape } from "../Simulateur/InformationsEtape";
+import {
+  AppartenancePaysUnionEuropeenne,
+  DesignationOperateurServicesEssentiels,
+  TrancheBilanFinancier,
+  TrancheChiffreAffaire,
+  TrancheNombreEmployes,
+  TypeEntitePublique,
+  TypeStructure,
+} from "../Simulateur/ChampsSimulateur.definitions";
+import { SecteurActivite } from "../Simulateur/SecteurActivite.definitions";
+import { SousSecteurActivite } from "../Simulateur/SousSecteurActivite.definitions";
+import { Activite } from "../Simulateur/Activite.definitions";
+
+export type EtatQuestionnaire = {
+  etapeCourante: TypeEtape;
+  designationOperateurServicesEssentiels: DesignationOperateurServicesEssentiels[];
+  appartenancePaysUnionEuropeenne: AppartenancePaysUnionEuropeenne[];
+  typeStructure: TypeStructure[];
+  trancheNombreEmployes: TrancheNombreEmployes[];
+  trancheChiffreAffaire: TrancheChiffreAffaire[];
+  trancheBilanFinancier: TrancheBilanFinancier[];
+  secteurActivite: SecteurActivite[];
+  sousSecteurActivite: SousSecteurActivite[];
+  activites: Activite[];
+  typeEntitePublique: TypeEntitePublique[];
+  localisationFournitureServicesNumeriques: AppartenancePaysUnionEuropeenne[];
+  paysDecisionsCyber: AppartenancePaysUnionEuropeenne[];
+  paysOperationsCyber: AppartenancePaysUnionEuropeenne[];
+  paysPlusGrandNombreSalaries: AppartenancePaysUnionEuropeenne[];
+};
+
+export const EtatQuestionnaireVide: EtatQuestionnaire = {
+  etapeCourante: "prealable",
+  designationOperateurServicesEssentiels: [],
+  appartenancePaysUnionEuropeenne: [],
+  typeStructure: [],
+  trancheNombreEmployes: [],
+  trancheChiffreAffaire: [],
+  trancheBilanFinancier: [],
+  secteurActivite: [],
+  sousSecteurActivite: [],
+  activites: [],
+  typeEntitePublique: [],
+  localisationFournitureServicesNumeriques: [],
+  paysDecisionsCyber: [],
+  paysOperationsCyber: [],
+  paysPlusGrandNombreSalaries: [],
+};
