+++ commun/core/src/Domain/Questionnaire/regles/RegleTaille.ts
@@ -1,9 +1,9 @@
-import { EtatQuestionnaire } from "../../reducerQuestionnaire.ts";
-import { estValeurVide, Regle } from "../Specifications.ts";
-import { UnionPetitMoyenGrand } from "../../../../../commun/core/src/Domain/Simulateur/ChampsSimulateur.definitions.ts";
-import { contientUnParmi } from "../../../../../commun/utils/services/commun.predicats.ts";
-import { SpecificationTexte } from "../FormatDesSpecificationsCSV.ts";
-import { ErreurLectureDeRegle } from "./ErreurLectureDeRegle.ts";
+import { estValeurVide, Regle } from "../Specifications";
+import { UnionPetitMoyenGrand } from "../../Simulateur/ChampsSimulateur.definitions";
+import { contientUnParmi } from "../../../../../utils/services/commun.predicats";
+import { ErreurLectureDeRegle } from "./ErreurLectureDeRegle";
+import { EtatQuestionnaire } from "../EtatQuestionnaire";
+import { SpecificationTexte } from "../FormatDesSpecificationsCSV";
 
 export class RegleTaille implements Regle {
   constructor(private readonly tailleAttendue: UnionPetitMoyenGrand) {}
